Add memory_resource_base
