+++ libcudacxx/include/cuda/__memory_pool/device_memory_pool.h
@@ -25,6 +25,7 @@
 
 #  include <cuda/__memory_pool/memory_pool_base.h>
 #  include <cuda/__memory_resource/get_property.h>
+#  include <cuda/__memory_resource/memory_resource_base.h>
 #  include <cuda/__memory_resource/properties.h>
 #  include <cuda/__runtime/api_wrapper.h>
 #  include <cuda/std/__concepts/concept_macros.h>
@@ -60,7 +61,9 @@ _CCCL_DIAG_SUPPRESS_CLANG("-Wmissing-braces")
 //!    exceeds the lifetime of the ``device_memory_pool_ref``.
 //!
 //! @endrst
-class device_memory_pool_ref : public __memory_pool_base
+class device_memory_pool_ref
+    : public __memory_pool_base
+    , public ::cuda::mr::memory_resource_base<device_memory_pool_ref>
 {
 public:
   //! @brief  Constructs the device_memory_pool_ref from a \c cudaMemPool_t.
